requirements.txt
attrs==25.1.0
blurhash==1.1.4
boost==0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
cmake==3.29.2
decorator==5.1.1
exceptiongroup==1.2.2
greenlet==3.0.3
h11==0.14.0
idna==3.10
Js2Py==0.74
Mastodon.py==1.8.1
numpy==2.0.2
outcome==1.3.0.post0
pandas==2.2.3
pycparser==2.22
pycryptodome==3.21.0
PyExecJS==1.5.1
pyjsparser==2.7.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-magic-bin==0.4.14
pytz==2024.2
requests==2.32.3
rsqlalchemy==0.0.1
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
